Implement pureComputed, and add unittests.
pureComputed() is similar to computed(), except that it only subscribes
to dependencies while it itself has subscribers (so for pure functions
it is suitable and essentially identical).

One advantage is that a pureComputed with no references to it can be
garbage-collected, whereas a computed() can be garbage-collected only
when its dependencies have no references either.